fix(core): declare 46 orphaned source modules in lib.rs and fix compilation
Add pub mod declarations for 46 source files that existed but weren't
compiled. Includes proper feature gates (subprocess-bridge for mission/tx/
planner/shadow/canary modules, vendored for vendored_migration_map,
web for web_framework). Adds MissionAgentAvailability enum and
MissionAgentCapabilityProfile struct to plan.rs (required by
planner_features and mission_loop). Fixes safe_channel Drop impl
Send bound mismatch and several pre-existing test compilation issues
(MuxHealthSample missing fields, RuleDef field renames, DecisionEvent
field changes).

+++ crates/frankenterm-core/src/planner_features.rs
@@ -690,7 +690,7 @@ pub fn solve_assignments(
             )
         })
         .map(|a| {
-            let cap = a.effective_capacity().saturating_sub(a.current_load);
+            let cap = a.effective_capacity().saturating_sub(a.current_load) as u32;
             (a.agent_id.clone(), cap)
         })
         .collect();
